refactor: extract transcription logic into transcription service
